+++ ui/pages/Project/components/EditDatasetButton.test.js
@@ -10,3 +10,12 @@ configure({ adapter: new Adapter() })
 test('shallow-render without crashing', () => {
   shallow(<EditDatasetsButton user={STATE_WITH_2_FAMILIES.user} />)
 })
+
+test('shallow-render load workspace data', () => {
+  shallow(
+    <EditDatasetsButton
+      project={STATE_WITH_2_FAMILIES.projectsByGuid.R0237_1000_genomes_demo}
+      user={STATE_WITH_2_FAMILIES.user}
+    />,
+  )
+})