+++ ui/pages/Report/Report.jsx
@@ -9,6 +9,7 @@ import { Error404, Error401 } from 'shared/components/page/Errors'
 import Anvil from './components/Anvil'
 import CustomSearch from './components/CustomSearch'
 import DiscoverySheet from './components/DiscoverySheet'
+import Gregor from './components/Gregor'
 import SampleMetadata from './components/SampleMetadata'
 import SeqrStats from './components/SeqrStats'
 
@@ -16,6 +17,7 @@ export const REPORT_PAGES = [
   { path: 'anvil', component: Anvil },
   { path: 'custom_search', params: '/:searchHash?', component: CustomSearch },
   { path: 'discovery_sheet', params: '/:projectGuid?', component: DiscoverySheet },
+  { path: 'gregor', component: Gregor },
   { path: 'sample_metadata', params: '/:projectGuid?', component: SampleMetadata },
   { path: 'seqr_stats', component: SeqrStats },
 ]